#111b Tickets #113 und #115, Umstellung der RestApi auf Contribution

+++ api/src/main/java/org/jqassistant/tooling/dashboard/api/dto/ContributionDTO.java
@@ -7,6 +7,6 @@ import lombok.*;
 @Setter
 @ToString
 public class ContributionDTO {
-    private int commits;
+    private long commits;
     private ContributorDTO contributor;
 }
